Update 0.0.47 Sick Nasty
My boyfriend called the game "Sick Nasty" so I'm calling this update that

Bug fixed a bunch of things and added minor changes all around to clean up the game and show off the demo.  Created a bunch of levels and added them to the Debug Rooms folder.

I'm gonna try working on creating ragdoll effects for the characters if they die.  I also changed the resolution of the game from 480x270 to 640x360

What I should work on is the Camera, Weapons in the Inventory, and Command Mode UI layout